Moved get_server_type to backends
