+++ dash/backends/__init__.py
@@ -1,7 +1,11 @@
+from .base_server import BaseDashServer, RequestAdapter
+
 from typing import Literal
 import importlib
 
-from .base_server import BaseDashServer, RequestAdapter
+
+request_adapter: RequestAdapter
+backend: BaseDashServer
 
 
 _backend_imports = {
@@ -9,9 +13,6 @@ _backend_imports = {
     "fastapi": ("dash.backends._fastapi", "FastAPIDashServer", "FastAPIRequestAdapter"),
     "quart": ("dash.backends._quart", "QuartDashServer", "QuartRequestAdapter"),
 }
-
-request_adapter: RequestAdapter
-backend: BaseDashServer
 
 
 def get_backend(name: Literal["flask", "fastapi", "quart"]):
@@ -33,8 +34,49 @@ def get_backend(name: Literal["flask", "fastapi", "quart"]):
         ) from e
 
 
+def _is_flask_instance(obj):
+    try:
+        # pylint: disable=import-outside-toplevel
+        from flask import Flask
+
+        return isinstance(obj, Flask)
+    except ImportError:
+        return False
+
+
+def _is_fastapi_instance(obj):
+    try:
+        # pylint: disable=import-outside-toplevel
+        from fastapi import FastAPI
+
+        return isinstance(obj, FastAPI)
+    except ImportError:
+        return False
+
+
+def _is_quart_instance(obj):
+    try:
+        # pylint: disable=import-outside-toplevel
+        from quart import Quart
+
+        return isinstance(obj, Quart)
+    except ImportError:
+        return False
+
+
+def get_server_type(server):
+    if _is_flask_instance(server):
+        return "flask"
+    if _is_quart_instance(server):
+        return "quart"
+    if _is_fastapi_instance(server):
+        return "fastapi"
+    raise ValueError("Invalid backend argument")
+
+
 __all__ = [
     "get_backend",
     "request_adapter",
     "backend",
+    "get_server_type",
 ]
